// web/src/api/generated/models/SearchResult.ts
/* tslint:disable */
/* eslint-disable */
/**
 * SideBySide Next
 * Application Core. OpenAPI is the authoritative contract.
 *
 * The version of the OpenAPI document: 0.1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
import type { SearchKind } from './SearchKind';
import {
    SearchKindFromJSON,
    SearchKindFromJSONTyped,
    SearchKindToJSON,
    SearchKindToJSONTyped,
} from './SearchKind';
import type { SearchScope } from './SearchScope';
import {
    SearchScopeFromJSON,
    SearchScopeFromJSONTyped,
    SearchScopeToJSON,
    SearchScopeToJSONTyped,
} from './SearchScope';

/**
 * 
 * @export
 * @interface SearchResult
 */
export interface SearchResult {
    /**
     * 
     * @type {string}
     * @memberof SearchResult
     */
    excerpt: string | null;
    /**
     * 
     * @type {string}
     * @memberof SearchResult
     */
    id: string;
    /**
     * 
     * @type {Date}
     * @memberof SearchResult
     */
    occurredOn: Date | null;
    /**
     * 
     * @type {string}
     * @memberof SearchResult
     */
    parentId: string | null;
    /**
     * 
     * @type {SearchScope}
     * @memberof SearchResult
     */
    scope: SearchScope;
    /**
     * 
     * @type {string}
     * @memberof SearchResult
     */
    title: string | null;
    /**
     * 
     * @type {SearchKind}
     * @memberof SearchResult
     */
    type: SearchKind;
}



/**
 * Check if a given object implements the SearchResult interface.
 */
export function instanceOfSearchResult(value: object): value is SearchResult {
    if (!('excerpt' in value) || value['excerpt'] === undefined) return false;
    if (!('id' in value) || value['id'] === undefined) return false;
    if (!('occurredOn' in value) || value['occurredOn'] === undefined) return false;
    if (!('parentId' in value) || value['parentId'] === undefined) return false;
    if (!('scope' in value) || value['scope'] === undefined) return false;
    if (!('title' in value) || value['title'] === undefined) return false;
    if (!('type' in value) || value['type'] === undefined) return false;
    return true;
}

export function SearchResultFromJSON(json: any): SearchResult {
    return SearchResultFromJSONTyped(json, false);
}

export function SearchResultFromJSONTyped(json: any, ignoreDiscriminator: boolean): SearchResult {
    if (json == null) {
        return json;
    }
    return {
        
        'excerpt': json['excerpt'],
        'id': json['id'],
        'occurredOn': (json['occurredOn'] == null ? null : new Date(json['occurredOn'])),
        'parentId': json['parentId'],
        'scope': SearchScopeFromJSON(json['scope']),
        'title': json['title'],
        'type': SearchKindFromJSON(json['type']),
    };
}

export function SearchResultToJSON(json: any): SearchResult {
    return SearchResultToJSONTyped(json, false);
}

export function SearchResultToJSONTyped(value?: SearchResult | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'excerpt': value['excerpt'],
        'id': value['id'],
        'occurredOn': value['occurredOn'] == null ? value['occurredOn'] : value['occurredOn'].toISOString().substring(0,10),
        'parentId': value['parentId'],
        'scope': SearchScopeToJSON(value['scope']),
        'title': value['title'],
        'type': SearchKindToJSON(value['type']),
    };
}